fixed problems with  http://issues.infrae.com/silva/issue14 i.e. encoding stuff with Document.store_xml
In the end i found out that the client side sent an xml-document
which was encoded differently in relation to what the xml-encoding
attribute stated.  The other problem was that set_title and
manage_edit have different expecations for their arguments
(set_title wants unicode, manage_edit wants utf8)

I hope this fixes the issues (on the server-side).