- nlist support
- lots o new unit tests

- improved list support

this was a major piece of work because the html list elements
are heavily overloaded with silva-counterparts.  As we can't
stick any meta-data to the html (is stripped out by RealObjects)
there is a lot of detection/correction and conversion stuff
for getting list, dlist, nlist and all of their types from
an 'ul' tag.