add dummy security trigger.
This is necessary to work aroud a stupid bug with some broken mozilla version,
which sometimes stop to send authorization headers, and wait for the server
to send the 401, then resending the data.
 As the save scripts only operate on the xml nodes, it seems Zope only
realizes the missing unauthorization when commiting the changes to the ZODB;
in this case there has been already sent a redirect status code, and
the browser just redirects instead of senting the auth headers.

 Hm, this is actually a hypothesis of mine, I admit. What I actually see
is that sometimes I push the "save" button and simply nothing seems
to happen. With this stupid fix, it works.