add a keyword parameter 'view_type' to the preview/view methods. This allows to register other public views (e.g. 'short','feature', etc) and call them via the "content.html" (or whatever).
 This may be useful for the News, where the News viewer does not have
to depend on the type of the news items; instead it semds them all a
'view(view_type="summary")' and the news items render a summary view
(if registered, of course). Actually Marc asked be how to do this,
thus I check in the little patch.

 As the keyword argument has the default value "public",
this should be fully backwards compatible.