Initial conversion module and tests.
This is currently not used anywhere in silva.
I'll go ahead and try to adapt my edit-on-pro
stuff with the simple 'convert' API.  Don't
know if this should go in the silva-release
next week.  You could say this is a 'new'
feature although users would not see it in any way.