First stab at title policies/index document policies.
There is now a new silva-publication metadata set that only
exists on publications. This set contains only a single field,
where the user (should be Chief Editor and up) can select what
kind of index should be created automatically for a particular
publication. This can be Silva Document, but can also be a SimpleContent
(as soon as I enabled this :), which is a trivial content object
with a non-versioned title. If SimpleContent is added, the effect
for the user will be like in Silva 0.9.1; non-versioned titles
that can be changed by changing the folder title.

Later on the plan is to create another index policy object.
This object will select the first actual content object and the folder
and effectively ghost to it. It will maintain its own title however.