Made SimpleContent behave properly with a title and allow this to be set as an index policy in publications. This means the promise in the previous checkin message has been fulfilled.
SimpleContent is of course not a very useful policy at present. It
could be made useful by automatically displaying a TOC in its
public view (though it would need to be renamed then to TOCContent or
something along those lines). We also need the DelegateContent.

I am tempted to call this 'TOCIndex' and 'DelegateIndex' but unfortunately
the word 'index' is far too heavily overloaded. TOCFolderContent
and DelegateFolderContent perhaps.. These would not be addable by
ordinary users anyway.