Renamed testServiceLayouts to test_ServiceLayouts to follow pattern.
Renamed testSilvaTestCase to skel.py. It contains example code for a
typical test module.

Got rid of base.py and security.py; they were not in use anymore.
test_all.py can also go away in favor of runalltests.py

testSkeletonSilvaTest.py was also superfluous; it did almost the same
as testSilvaTestCase (function as a skeleton) with only very minor
differences. Got rid of it.

I added a README.txt describing how to run the tests and how to
use the test framework.

(note, not all test modules run yet at the time of this checkin)