fix for one part of issue 526:
If a ghost is published, the Editor screen displays only the option to create a
new version. There is no information about the ghosted object. To find out what
it references, a user has to make a new version. We should at least show the
ghosted object's location (rendered as a link to its tab_edit).