Improved upgrade code during EUR data upgrades:
o Replaced (some of the) LOG entries with print statements. Somehow
the LOG entries go to nowhere when invoking the upgrade from the
commandline e.g. using the silva_maintenance scripts. This should be
figured out later though as LOG entries are more elegant.

o Set the "RefreshAll" upgrade earlier in the upgraders sequence again.
Somwhoe, refresh after all upgrading is done, breaks some of the upgraders.
This should be investigated later to maybe, but for EUR this now works.

o It seemed that the 'check_reserved_ids" upgrader was called for all
meta_types - fixed it by registering the upgrader only for the
'Silva Root' object.

o Followed Clemens' suggestion to make the upgradeTree fdo a depth-first
upgrade instead of a breadth-first upgrade. This keeps the length of the
upgrade queue to a smaller number.

o Numerous smaller and larger fixes in the upgraders