fixed bug which caused the listing position numbers to have a weird order (0, 2, 3, 4...) when an index was present
my god, this is ugly code; it's starting to look like perl

and, since I'm in bitching mode, I can't resist pointing out that this
ugly code is necessary to 'code around' the index, something we have
to do all time, in product code too (is it an index? do something
special; is it not? do normal) and life and code will be a lot simpler
when there isn't an index, someday, sigh, sssss