Adding fullmedia import/export
* For chief editors and managers an extra form is exposed on the export page
to export a fullmedia archive. Apart from the 'include subpublications'
option, an option has been added to export only the last (possibly
unpublished) version. If not checked, all versions will be exported.

* The zip import functionality has been changed to test if the imported
archive is a fullmedia archive - which now is done simply by checking if the
archive contains a 'silva.xml' file. This should be changed in the future;
fullmedia archives should have a manifest file which uniquely identifies them.

* The 'slv' extension for silva xml has been dropped in favour of plain 'xml'