First merge-in of data extracted from forms.
Unfortunately the message ids are fairly obscure, based on fields in the
form, not what you see on the screen. We'll have to live with this for
now though. Systematic naming of the message ids should ease the pain.

Unfortunately I cannot extract message ids from the metadata XML yet,
as this uses a different serialization than the one Formulator uses...

I've only included a single actual dutch translation of a single
description (for id of Silva File), but it's a start.