Re-ran i18nextract and i18nmerge tools to get the translations updated, while doing so I noticed that there were duplicate messageids, which made that the merging failed. I removed the duplicates, everything seems to work now.
Do note that the duplicates seem to suggest that someone has *manually*
changed msgids in the .po files (further investigation pointed to changes
in the .pot files, too), which should *never* happen! Always take the
following steps to keep the translations up-to-date:

1. Change strings in code/views
2. Run i18nextract to generate the .pot file
3. Run the i18nmerge tool to update the .po files
4. Add/edit the translations (obviously this will have to happen manually)

Changing the .po or .pot files directly can and will lead to broken .po files
and/or data loss.