Remove Silva-the-Product's own view lookup mechanisms for default views in favour of SilvaLayout.
Although we had the two working together when we committed this,
SilvaLayout and Silva didn't work together now.  And we found that on
the view side, there's too much conceptual overlap to keep the two
"ways of doing it".

Support for edit views was taken out, too.  There's code in there to
maybe pick this up later, see FrankenViewAttribute.